feat(scanner): `-wl-import` option, better doc comment placement
The `-wl-import:<string>` option should be used to set the import used
for the core Wayland package. It is not

The doc comments for requests are now placed above their respective
procedures and not the opcode constants.
It doesn't really make sense for the documentation to be on the opcodes
and not the procedures, as opcodes are rarely if ever actually
referenced in user code.
In fact, why do the opcode constants even exist? They're only ever used
once and could very reasonably be inlined (as far as I see), but that's
more than I want to do for this.